+++ NEWS.md
@@ -1,10 +1,14 @@
 ## Version 0.1.5
 
----
-
  - added vignette
  - removeLazyData
  - DESCRIPTION updaates
  - replacing url_exists with RCurl::url.exists
  - added NEWS.md
  
+## Version 0.1.6
+
+ - bug fix in fetch_kegg
+ - removed `min_count` and `fdr_limit` arguments from `functional_enrichment`; filtering can be done afterwards
+ - added a small Shiny app as an example of `fenr`
+ - updates to vignette